Renaming the file DeleteClientModel
